Connection fix (hopefully)

+++ commands/video_player.js
@@ -16,8 +16,7 @@ module.exports = {
         const connection = await voice_channel.join();
         connection.voice.setSelfDeaf(true);
         song_queue.connection = connection;
-
-
+        
         const videoPlayer = async(guild, song) => {
             
             const video_finder = async(query) => {
@@ -42,30 +41,38 @@ module.exports = {
 
             let stream = ytdlDisc(song.url, {
                 filter: "audioonly",
-                quality: "highestaudio",
                 opusEncoded: true,
             });
 
-            song_queue.connection.play(stream, {seek: 0, volume: 0.45, type: "opus", bitrate: 'auto', fec: true})
-            .on('start' , () => {
-                const playingEmbed = new Discord.MessageEmbed() 
-                .setColor("#8deeee")
-                .setTitle("Now Playing")
-                .setDescription(`**${song.title}**`);
 
-                song_queue.text_channel.send(playingEmbed);
-            })
-            .on('finish' , () => {
-                if(song_queue.songs.length !== 1) {
-                    song_queue.songs.shift();
-                    song_queue.connection.dispatcher.end();
-                    return videoPlayer(guild, song_queue.songs[0]);
-                } else {
-                    queue.removeQueueID(guild.id);
-                    timeoutID = setTimeout(() => {voice_channel.leave()}, 10000) // Dies in 3 Minute for heroku not to kill the connection and cause a bug to happen
-                }
-            });
-            
+            if(connection != null) {
+                song_queue.connection.play(stream, {seek: 0, volume: 0.45, type: "opus", bitrate: 'auto', fec: true})
+                .on('start' , () => {
+                    const playingEmbed = new Discord.MessageEmbed() 
+                    .setColor("#8deeee")
+                    .setTitle("Now Playing")
+                    .setDescription(`**${song.title}**`);
+
+                    song_queue.text_channel.send(playingEmbed);
+                })
+                .on('finish' , () => {
+                    if(song_queue.songs.length !== 1) {
+                        song_queue.songs.shift();
+                        return videoPlayer(guild, song_queue.songs[0]);
+                    } else {
+                        queue.removeQueueID(guild.id);
+                        timeoutID = setTimeout(() => {voice_channel.leave()}, 10000) // Dies in 3 Minute for heroku not to kill the connection and cause a bug to happen
+                    }
+                });
+            } else {
+                const connection = await voice_channel.join();
+                connection.voice.setSelfDeaf(true);
+                song_queue.connection = connection;
+                return videoPlayer(guild, song_queue.songs[0]);
+            }
+
+
+
         }
 
         videoPlayer(guild, song);
